refactor: multi->chess 패키지명 변경
