+++ src/main/java/com/example/chessserver/chess/domain/Piece.java
@@ -1,4 +1,4 @@
-package com.example.chessserver.multi.domain;
+package com.example.chessserver.chess.domain;
 
 public enum Piece {
     KING,
