+++ src/main/java/com/example/chessserver/chess/domain/Player.java
@@ -1,4 +1,4 @@
-package com.example.chessserver.multi.domain;
+package com.example.chessserver.chess.domain;
 
 import lombok.Getter;
