+++ src/main/java/com/example/chessserver/chess/exception/NoContentException.java
@@ -1,4 +1,4 @@
-package com.example.chessserver.multi.exception;
+package com.example.chessserver.chess.exception;
 
 public class NoContentException extends RuntimeException{
