+++ src/main/java/com/example/chessserver/chess/service/TestService.java
@@ -1,6 +1,6 @@
-package com.example.chessserver.multi.service;
+package com.example.chessserver.chess.service;
 
-import com.example.chessserver.multi.repository.GameRepository;
+import com.example.chessserver.chess.repository.GameRepository;
 import com.example.chessserver.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
